Create Project 1- Java
Create a class Student which contains firstName, lastName, studentID (4 digit String) and implements the Comparable type. The class includes a default constructor (null for names, 0 for ID) and a constructor that creates a Student with parameters firstName, lastName, and StudentID. The class will also have “setters” and “getters” (accessors and mutators) for all 3 data fields. The class will also include a method .equals(Student other) that will check the StudentID, a .toString  that will format the Student output in any way you choose, and a .compareTo (Student other)  that will be used to sort the students by lastName (and then firstName if lastName is the same).