Create Sentiment Analysis
This is a basic sentiment analyzer that scores each word of the input text as positive +1, neutral 0, or negative -1. The sum of all the words in the text becomes the sentiment score. Text with an overall score >0 will be considered positive, likewise, text will an overall score of <0 will be considered negative.